Refactor v1/v2 runner components
