+++ fractal_server/app/runner/v1/_local/__init__.py
@@ -25,9 +25,9 @@ from typing import Optional
 
 from ....models import Workflow  # FIXME: this is v1 specific
 from ...executors.local.executor import FractalThreadPoolExecutor
+from ...set_start_and_last_task_index import set_start_and_last_task_index
 from .._common import execute_tasks  # FIXME: this is v1 specific
-from ..common import async_wrap
-from ..common import set_start_and_last_task_index
+from ..async_wrap import async_wrap
 from ..common import TaskParameters  # FIXME: this is v1 specific
 from ._submit_setup import _local_submit_setup
