+++ tests/01_unit/test_unit_set_start_and_last_task_index.py
@@ -4,7 +4,9 @@ from typing import Union
 import pytest
 from devtools import debug
 
-from fractal_server.app.runner.v1.common import set_start_and_last_task_index
+from fractal_server.app.runner.set_start_and_last_task_index import (
+    set_start_and_last_task_index,
+)
 
 
 cases = [
