+++ tests/fixtures_server_v2.py
@@ -8,7 +8,7 @@ from fractal_server.app.models.v2 import JobV2
 from fractal_server.app.models.v2 import ProjectV2
 from fractal_server.app.models.v2 import WorkflowV2
 from fractal_server.app.routes.api.v2.apply import _encode_as_utc
-from fractal_server.app.runner.v1.common import (
+from fractal_server.app.runner.set_start_and_last_task_index import (
     set_start_and_last_task_index,  # FIXME v2
 )
